eval/requirements.txt
tqdm
openai
tenacity
python-dotenv
fastapi
uvicorn